requirements.txt
attrs==19.3.0
-e git+https://github.com/lokedede/aws-fargate-cdk.git@ee239d1e73c9b55ba6c6ea288819a2d956813ff1#egg=aws_cdk_provision
aws-cdk.assets==1.52.0
aws-cdk.aws-apigateway==1.52.0
aws-cdk.aws-applicationautoscaling==1.52.0
aws-cdk.aws-autoscaling==1.52.0
aws-cdk.aws-autoscaling-common==1.52.0
aws-cdk.aws-autoscaling-hooktargets==1.52.0
aws-cdk.aws-batch==1.52.0
aws-cdk.aws-certificatemanager==1.52.0
aws-cdk.aws-cloudformation==1.52.0
aws-cdk.aws-cloudfront==1.52.0
aws-cdk.aws-cloudwatch==1.52.0
aws-cdk.aws-codebuild==1.52.0
aws-cdk.aws-codecommit==1.52.0
aws-cdk.aws-codeguruprofiler==1.52.0
aws-cdk.aws-codepipeline==1.52.0
aws-cdk.aws-cognito==1.52.0
aws-cdk.aws-ec2==1.52.0
aws-cdk.aws-ecr==1.52.0
aws-cdk.aws-ecr-assets==1.52.0
aws-cdk.aws-ecs==1.52.0
aws-cdk.aws-ecs-patterns==1.52.0
aws-cdk.aws-efs==1.52.0
aws-cdk.aws-elasticloadbalancing==1.52.0
aws-cdk.aws-elasticloadbalancingv2==1.52.0
aws-cdk.aws-events==1.52.0
aws-cdk.aws-events-targets==1.52.0
aws-cdk.aws-iam==1.52.0
aws-cdk.aws-kinesis==1.52.0
aws-cdk.aws-kms==1.52.0
aws-cdk.aws-lambda==1.52.0
aws-cdk.aws-logs==1.52.0
aws-cdk.aws-route53==1.52.0
aws-cdk.aws-route53-targets==1.52.0
aws-cdk.aws-s3==1.52.0
aws-cdk.aws-s3-assets==1.52.0
aws-cdk.aws-sam==1.52.0
aws-cdk.aws-secretsmanager==1.52.0
aws-cdk.aws-servicediscovery==1.52.0
aws-cdk.aws-sns==1.52.0
aws-cdk.aws-sns-subscriptions==1.52.0
aws-cdk.aws-sqs==1.52.0
aws-cdk.aws-ssm==1.52.0
aws-cdk.aws-stepfunctions==1.52.0
aws-cdk.cloud-assembly-schema==1.52.0
aws-cdk.core==1.52.0
aws-cdk.custom-resources==1.52.0
aws-cdk.cx-api==1.52.0
aws-cdk.region-info==1.52.0
cattrs==1.0.0
constructs==3.0.4
importlib-resources==3.0.0
jsii==1.9.0
pkg-resources==0.0.0
publication==0.0.3
python-dateutil==2.8.1
six==1.15.0
typing-extensions==3.7.4.2
zipp==3.1.0
